+++ bullet.rb
@@ -1,47 +1,47 @@
 class Bullet < Sprite
-    def initialize(x,y)
-        bullet = Image.new(100,20,C_RED)
-        super(x,y,bullet)
-        @speed = 5
-    end
+  def initialize(x, y)
+    bullet = Image.new(100, 20, C_RED)
+    super(x, y, bullet)
+    @speed = 5
+  end
 
-    def update
-        self.x += @speed
-        #self.y += @speed
-        if self.x > 800
-            self.vanish
-        end
+  def update
+    self.x += @speed
+    #self.y += @speed
+    if self.x > 800
+      self.vanish
     end
+  end
 
-    def shot
-        self.vanish
-    end
+  def shot
+    self.vanish
+  end
 end
 
 class Bullets
-    MAX_BALLET = 5
+  MAX_BALLET = 5
 
-    def initialize
-        @box = []
-        @count = 0
-    end
+  def initialize
+    @box = []
+    @count = 0
+  end
 
-    def update(enemies,ways)
-        Sprite.check(@box,enemies)
-        Sprite.check(@box,ways)
-        Sprite.update(@box)
-        Sprite.clean(@box)
-        
-        if Input.key_push?(K_SPACE) && @box.size < MAX_BALLET
-            @box << Bullet.new($x, $y)
-        end
-    end
+  def update(enemies, ways)
+    Sprite.check(@box, enemies)
+    Sprite.check(@box, ways)
+    Sprite.update(@box)
+    Sprite.clean(@box)
 
-    def draw
-        Sprite.draw(@box)
+    if Input.key_push?(K_SPACE) && @box.size < MAX_BALLET
+      @box << Bullet.new($x, $y)
     end
+  end
 
-    def getCount
-        @count
-    end
+  def draw
+    Sprite.draw(@box)
+  end
+
+  def getCount
+    @count
+  end
 end
