+++ cloud.rb
@@ -3,15 +3,8 @@ class Cloud < Sprite
         @speed = 8
         image =Image.load('images/kumo.png')
         image.set_color_key(C_GREEN)
-        #super(x,y,image)
-        #@cloud1 = Sprite.new(50,15,@image)
-        #@cloud2 = Sprite.new(200,15,@image)
-        #@cloud3 = Sprite.new(350,15,@image)
-        #@cloud4 = Sprite.new(500,15,@image)
-        #@cloud5 = Sprite.new(650,15,@image)
         y = 15r
         super(x,y,image)
-        # @enemies << Enemy.new(800, @enemyPlace[rand(0..1)])
     end
 
     def update
@@ -20,10 +13,6 @@ class Cloud < Sprite
             self.vanish
         end
     end
-
-    # def getSprite
-    #     @clouds
-    # end
     
     def shot
         puts "hit!"
@@ -31,7 +20,6 @@ class Cloud < Sprite
 end
 
 class Clouds
-
     MAX_CLOUD = 4
 
     def initialize
